+++ color_diary_project/color_diary/models/__init__.py
@@ -1,0 +1,3 @@
+from .user import *
+from .color import *
+from .diary import *
