+++ color_diary_project/color_diary/tests/__init__.py
@@ -1,0 +1,6 @@
+from .admins import *
+from .constant import *
+from .fields import *
+from .models import *
+from .signals import *
+from .views import *
